+++ ak_library_management/models/library_members.py
@@ -1,13 +1,16 @@
 # -*- coding: utf-8 -*-
 
 from odoo import models,fields,api,_
+from odoo.exceptions import ValidationError
 
 
 class LibraryMembers(models.Model):
     """This class is storing fields of library member for the record."""
     _name = 'library.members'
     _description = 'Library Members'
+    _rec_name = 'lib_member_id'
 
+    lib_member_id = fields.Many2one(comodel_name='res.partner', string="Library Member Id")
     name = fields.Char(string='Member Name', required=True)
     email = fields.Char(string='Email ID')
     phone = fields.Char(string='Contact Number')
@@ -16,9 +19,30 @@ class LibraryMembers(models.Model):
 
     @api.model_create_multi
     def create(self, vals_list):
-        """to generate the sequence for the members of library"""
+        """to generate the unique sequence for the members of library.
+        param: vals_list
+        return: True"""
         for vals in vals_list:
             if not vals.get('member_id') or vals['member_id'] == _('New'):
                 vals['member_id'] = (self.env['ir.sequence']
                                      .next_by_code('library.membership') or _('New'))
         return super().create(vals_list)
+
+    def send_renewal_mail(self):
+        """return the mail compose wizard for the renewal of membership.
+        Only manager or librarian can send the mail.
+        param: None
+        return: True"""
+        template = self.env.ref('ak_library_management.send_email_renewal_membership')
+        ctx = {'default_template_id':template.id}
+        if self.env.user.is_manager:
+            return {
+                'name': _('Renewal Membership'),
+                'type': 'ir.actions.act_window',
+                'view_mode': 'form',
+                'res_model': 'mail.compose.message',
+                'target': 'new',
+                'context': ctx,
+            }
+        raise ValidationError(_('You are not allowed to send email as you are not a manager.'))
+