+++ ak_library_management/models/library_template.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
 
-from odoo import models,fields,api
+from odoo import models, fields, api
 
 
 class Library(models.Model):
     """" This model gives the information about the library."""
     _name = 'library.library'
-    _inherit = ['mail.thread','mail.activity.mixin']
+    _inherit = ['mail.thread', 'mail.activity.mixin']
     _description = 'Library'
 
     name = fields.Char(string='Name', required=True)
@@ -27,13 +27,17 @@ class Library(models.Model):
 
     @api.depends('book_ids')
     def count_borrowed_book(self):
-        """function to count borrowed books"""
+        """count all the books borrowed by the customer.
+        param: None
+        return : True"""
         for rec in self:
             borrowed_books = rec.filtered(lambda reg: reg.book_ids.status == 'borrowed')
             self.count = len(borrowed_books)
 
     def action_count_borrowed_books(self):
-        """function of smart button name borrowed"""
+        """open the list of the borrowed books
+        param: None
+        return: True"""
         return {
             'type': 'ir.actions.act_window',
             'name': 'Borrowed Books List',